requirements.txt
flask>=2.0.0
    flask-sqlalchemy>=3.0.0
    flask-socketio>=5.3.0
    eventlet>=0.33.0
    python-socketio>=5.0.0
    flask-cors>=3.0.0
    ultralytics>=8.0.0
    opencv-python>=4.5.0
    numpy>=1.23.0
    torch>=2.0.0
    logging